# src/utils/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "Moovas Backend API",
    "contact": {
      "email": "thanhtutoo95@gmail.com"
    }
  },
  "host": "localhost:8000",
  "schemes": ["https", "http"],
...
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Returns user's data",
        "description": "",
        "operationId": "getsUser",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/getUserdata"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Updates user's data",
        "description": "",
        "operationId": "setsUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{login}/isLogin": {
      "get": {
        "tags": ["Users"],
        "summary": "Returns boolean that login exist",
        "description": "",
        "operationId": "isLogin",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "description": "",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "boolean",
            "schema": {
              "$ref": "#/definitions/isLogin"
            }
          }
        }
      }
    },
    "/api/users/{email}/isEmail": {
      "get": {
        "tags": ["Users"],
        "summary": "Returns boolean that email exist",
        "description": "",
        "operationId": "isEmail",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "boolean",
            "schema": {
              "$ref": "#/definitions/isEmail"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    },
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "getUserdata": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "lastSuccessfulLoggedDate": {
          "type": "string",
          "format": "date-time"
        },
        "lastPresentLoggedDate": {
          "type": "string",
          "format": "date-time"
        },
        "lastFailedLoggedDate": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "UpdateUser": {
      "type": "object",
      "properties": {
        "currencyId": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "isLogin": {
      "type": "object",
      "properties": {
        "isLogin": {
          "type": "boolean"
        }
      }
    },
    "isEmail": {
      "type": "object",
      "properties": {
        "isEmail": {
          "type": "boolean"
        }
      }
    },
...
  },
...
}